# documents/openapi.yaml
openapi: 3.0.3
info:
  title: Tijara API
  description: >
    REST API for Tijara, an e-commerce backend built with NestJS.
    A separate GraphQL API also exists for the Category module at
    POST /graphql (schema in src/schema.gql) and is not documented here.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
security: []
paths:
...
  /admin/brands:
    post:
      tags: [Brand - Admin]
      operationId: createBrand
      summary: Create a brand (optional logo upload)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateBrandDto"
                - type: object
                  properties:
                    logo:
                      type: string
                      format: binary
                      description: image/png, image/jpeg, image/webp, max 2MB
      responses:
        "201": { $ref: "#/components/responses/MessageOk" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/brands/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Brand - Admin]
      operationId: updateBrand
      summary: Update a brand (optional logo replace)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateBrandDto"
                - type: object
                  properties:
                    logo:
                      type: string
                      format: binary
      responses:
        "200": { $ref: "#/components/responses/MessageOk" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Brand - Admin]
      operationId: deleteBrand
      summary: Delete a brand
      security: [{ cookieAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/MessageOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- Category - Guest ----------------
...
  /admin/products:
    post:
      tags: [Product - Admin]
      operationId: createProduct
      summary: Create a product (up to 10 images)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateProductDto"
                - type: object
                  properties:
                    images:
                      type: array
                      maxItems: 10
                      items: { type: string, format: binary }
                      description: image/png, image/jpeg, image/webp, max 5MB each
      responses:
        "201": { $ref: "#/components/responses/MessageOk" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/products/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Product - Admin]
      operationId: updateProduct
      summary: Update a product (optional image replace)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateProductDto"
                - type: object
                  properties:
                    images:
                      type: array
                      maxItems: 10
                      items: { type: string, format: binary }
      responses:
        "200": { $ref: "#/components/responses/MessageOk" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Product - Admin]
      operationId: deleteProduct
      summary: Delete a product
      security: [{ cookieAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/MessageOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/products/{id}/stock:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Product - Admin]
      operationId: adjustProductStock
      summary: Adjust stock quantity (positive or negative delta)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdjustStockDto" }
      responses:
        "200": { $ref: "#/components/responses/MessageOk" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------- Coupon - Customer ----------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: Set by POST /auth/login. Required for all protected routes.
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Set by POST /auth/login. Used only by POST /auth/refresh-token.
  responses:
    MessageOk:
      description: Success
      content:
        application/json:
          schema: { $ref: "#/components/schemas/MessageResponse" }
...
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing/invalid/expired auth cookie
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated but role not permitted, or email not verified
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode: { type: integer, example: 400 }
        message: { type: string }
        error: { type: string }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
        payload: { type: object, additionalProperties: true }
...
    CreateBrandDto:
      type: object
      required: [name, categoryIds]
      properties:
        name: { type: string, minLength: 2, maxLength: 64 }
        categoryIds:
          type: array
          items: { type: string }
          minItems: 1
        description: { type: string, maxLength: 1024 }
    UpdateBrandDto:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 64 }
        categoryIds:
          type: array
          items: { type: string }
        description: { type: string, maxLength: 1024 }

    # ---- Cart DTOs ----
...
    CreateProductDto:
      type: object
      required: [name, price, categoryId, brandId]
      properties:
        name: { type: string, minLength: 2, maxLength: 128 }
        description: { type: string, maxLength: 2048 }
        price: { type: number, minimum: 0 }
        discountPercent: { type: number, minimum: 0, maximum: 100 }
        stock: { type: number, minimum: 0 }
        categoryId: { type: string }
        brandId: { type: string }
    UpdateProductDto:
      type: object
      description: Same as CreateProductDto minus `stock` (use the /stock endpoint instead)
      properties:
        name: { type: string, minLength: 2, maxLength: 128 }
        description: { type: string, maxLength: 2048 }
        price: { type: number, minimum: 0 }
        discountPercent: { type: number, minimum: 0, maximum: 100 }
        categoryId: { type: string }
        brandId: { type: string }
    AdjustStockDto:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          description: Positive or negative delta applied to current stock

    # ---- Coupon DTOs ----
...